Modificacion de la estructura del cliente
