+++ pom.xml
@@ -246,7 +246,7 @@
               </execution>
             </executions>
             <configuration>
-              <mainClass>com.psc06.client.ExampleClient</mainClass>
+              <mainClass>com.psc06.client.MainClient</mainClass>
               <arguments>
                 <argument>${server.hostname}</argument>
                 <argument>${server.port}</argument>
