+++ src/main/java/com/psc06/client/ClientServer.java
@@ -9,33 +9,19 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
-
 import com.psc06.pojo.UserStoryData;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class ExampleClient {
+public class ClientServer {
 
 	protected static final Logger logger = LogManager.getLogger();
-
-	// Userstory 1
-	private static final int id1 = 1;
-	private static final String userstory1 = "Crear cliente y servidor";
-	private static final int est1 = 5;
-	private static final int pb1 = 2;
-
-	// Userstory 2
-	private static final int id2 = 2;
-	private static final String userstory2 = "Crear pom";
-	private static final int est2 = 2;
-	private static final int pb2 = 7;
-
 
 	private Client client;
 	private WebTarget webTarget;
 
-	public ExampleClient(String hostname, String port) {
+	public ClientServer(String hostname, String port) {
 		client = ClientBuilder.newClient();
 		webTarget = client.target(String.format("http://%s:%s/rest/resource", hostname, port));
 	}
@@ -55,20 +41,5 @@ public class ExampleClient {
 		} else {
 			logger.info("User Story correctly created");
 		}
-	}
-
-	public static void main(String[] args) {
-		if (args.length != 2) {
-			logger.info("Use: java Client.Client [host] [port]");
-			System.exit(0);
-		}
-
-		String hostname = args[0];
-		String port = args[1];
-
-		ExampleClient conSer = new ExampleClient(hostname, port);
-		conSer.registerUserStory(id1, userstory1, est1, pb1);
-		conSer.registerUserStory(id2, userstory2, est2, pb2);
-		
 	}
 }
